requirements.txt
jinja2
decorator
flask
flask_mongoengine
flask_login
flask_babel
flask_cors
flask_jsglue
flask_mail
flask_menu
flask_breadcrumbs
flask_celeryext
flask_session
flask_swagger
wtforms_json
bleach
humanize
mongoengine_goodjson
pillow
lxml
passwordmeter
flask_admin
requests
# wtformsparsleyjs
openpyxl==2.3.5
xlsxwriter==0.9.2
simplejson
pytest
pytest-cov
pytest-instafail
pytest-sugar
python-coveralls
